added seedTest
