+++ script/seed.spec.js
@@ -1,6 +1,4 @@
 'use strict'
-/* global describe beforeEach it */
-
 const seed = require('./seed')
 
 describe('seed script', () => {
